Remove hardcoded irq numbers
This commit enables obtaining of interrupt numbers from DT instead of
values hardcoded in driver. It also describes new behaviour of the
driver in README and sample DTS.

I couldn't find the origins of irq numbers that were described in DTS,
but it is evident that the driver doesn't work with them. I have
replaced them with new ones that describe exactly 72-th and 73-th
interrupts.

git-svn-id: svn://svn.ddg/soc/branches/cpu/aes-crypto-api-iface@71305 b56743da-d8c8-0310-ae72-f310648a94a5