Move hardware info to a separate structure
This commit is a part of bigger refactoring, it moves duplicate fields
of `struct aes_priv` to a separate structure and instantiates that
structure twise. This was done to reduce code duplication and allow
later to get rid of `is_encrypt` flag being passed to many functions and
pass a pointer to `struct aes_priv_hwinfo` instead.

git-svn-id: svn://svn.ddg/soc/branches/cpu/aes-crypto-api-iface@71304 b56743da-d8c8-0310-ae72-f310648a94a5