Rewrite sg_feed_all to process both src and dst at once
This commit replaces two calls to sg_feed_all (the first one for src and
the second one for dst) with a single call that processes both src and
dst. This also changes the order in which descriptors are being fed: now
the src and dst descriptors are interspersed to allow FPGA to start
processing the data before all the descriptors are fed.