Use devres for allocating IRQs
It is done as a beginning of refactoring by moving allocation of all the
resources to devres. We actually do not take advantage of devres's
capability to free IRQs for us because we free IRQs implicitly to make
sure that no interrupt arrives at the time of freeing other resources.