Remove unnecessary wmb()
I couldn't find any information neither in comments nor in commit
messages on the purpose of that call to wmb(). Looks like it was added
just-in-case or copy-and pasted from some other project where it was
necessary. I see no reason in doing such a call so in this commit I
remove it for clarity.